답변 추천기능 구현

+++ src/main/java/com/exam/sbb/question/QuestionController.java
@@ -130,7 +130,7 @@ public class QuestionController {
 
   @PreAuthorize("isAuthenticated()")
   @GetMapping("/vote/{id}")
-  public String questionVote(Principal principal, @PathVariable("id") Integer id) {
+  public String questionVote(Principal principal, @PathVariable("id") Long id) {
     Question question = questionService.getQuestion(id);
     SiteUser siteUser = userService.getUser(principal.getName());
